requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
sqlalchemy-cockroachdb>=2.0.0
Werkzeug==3.0.1
beautifulsoup4==4.12.2
Authlib>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn==21.2.0
psycopg2-binary>=2.9.9
Flask-Compress>=1.14
